+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `input tag` component (Orange-OpenSource/ouds-ios#917)
 - Swift DocC tutorial (Orange-OpenSource/ouds-ios#919)
 
 ### Changed
@@ -27,7 +28,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Color` semantic tokens (tokens library v1.8.0) (Orange-OpenSource/ouds-ios#1070)
 - `Control item` component tokens (tokens library v1.8.0) (Orange-OpenSource/ouds-ios#1070)
 - Update `text input` to display error message if error, and display suffix/prefix even if placeholder empty (Orange-OpenSource/ouds-ios#1063)
-- `input tag` component (Orange-OpenSource/ouds-ios#917)
 
 ## [0.20.0](https://github.com/Orange-OpenSource/ouds-ios/compare/0.19.0...0.20.0) - 2025-10-15